a big reset - split out keycodes, layouts, and layers into separate files - just has basic includes
