+++ _keycodes.h
@@ -1,0 +1,23 @@
+#pragma once
+
+enum keycodes {
+  QWERTY = SAFE_RANGE,
+  CLMAKDH,
+  GAMING,
+
+  //These are only here to make the case statements unique in taphold.c
+  TH_LBRC,
+  TH_LABK,
+  TH_SQUO,
+  TH_DQUO,
+  TH_LPRN,
+  TH_PERC,
+  TH_COMM,
+  TH_DOT  
+};
+
+#define _DEFAULT_LAYER_1 FIRST_DEFAULT_LAYER
+#define _DEFAULT_LAYER_2 (FIRST_DEFAULT_LAYER + 1)
+#define _DEFAULT_LAYER_3 (FIRST_DEFAULT_LAYER + 2)
+
+#define ___x___ KC_NO
