+++ _layers.h
@@ -1,0 +1,14 @@
+#pragma once
+
+enum layers {
+  _QWERTY = 0,
+  FIRST_DEFAULT_LAYER = 0,  
+  _COLEMAK_DH,
+  _GAMING,
+  _NAVIGATION,
+  _SYMBOL,
+  _NUMBER,
+  _MOUSE,
+  _GAMENUMBER,
+  _CONFIG
+};
