+++ config.h
@@ -1,6 +1,8 @@
 #pragma once
+
 /*
 #include "layouts.h"
+
 #include "config/tapping_config.h"
