+++ rules.mk
@@ -1,6 +1,5 @@
 SRC += \
-	t4corun.c \
-	features/tapping1.c
+	t4corun.c
 
 # default settings across all keyboards
 # ?= means set the value if its not defined elsewhere
