+++ t4corun.h
@@ -1,2 +1,6 @@
 #pragma once
 #include QMK_KEYBOARD_H
+
+#include "_keycodes.h"
+#include "_layers.h"
+#include "_layouts.h"
